I implemented FireBase into my application and added such key aspects as: - FireBase Analytics - FireBase Authorization - FireBase DataBase -FireBase FireStore
Added the ability to authorize the user without re-authorization when re-entering the application if the user is registered.
The ability to post your jokes online, and display your posted jokes on your account. So, when you change your account, all the anecdotes from the network will be visible on the home page and your personal ones on your account.